Move NetworkTracePoller to a new file
This finishes the split of NetworkTraceHandler into separate DataSource
and Polling parts by moving the Polling part to its own file.

Despite being a large diff, all this change did was copy the
NetworkTraceHandler files to their NetworkTracePoller counterparts and
delete the irrelevant sections in each file. The actual content of the
classes and functions should be identical.

Bug: 246985031
Test: atest libnetworkstats_test
Change-Id: Ibc9d945658e89f969fa3d1551863ccd26fd51a78